Move masterwork info to masterwork.ts
