+++ src/app/inventory/store/masterwork.ts
@@ -1,0 +1,163 @@
+import { D2Item, DimSockets, DimMasterwork } from '../item-types';
+import { D2ManifestDefinitions } from 'app/destiny2/d2-definitions.service';
+import idx from 'idx';
+import { DamageType } from 'bungie-api-ts/destiny2';
+
+/**
+ * These are the utilities that deal with figuring out Masterwork info.
+ *
+ * This is called from within d2-item-factory.service.ts
+ */
+
+const resistanceMods = {
+  1546607977: DamageType.Kinetic,
+  1546607980: DamageType.Void,
+  1546607978: DamageType.Arc,
+  1546607979: DamageType.Thermal
+};
+
+/**
+ * This builds the masterwork info - this isn't whether an item is masterwork, but instead what
+ * "type" of masterwork it is, what the kill tracker value is, etc. Exotic weapons can start having
+ * kill trackers before they're masterworked.
+ */
+export function buildMasterwork(
+  createdItem: D2Item,
+  defs: D2ManifestDefinitions
+): DimMasterwork | null {
+  if (!createdItem.sockets) {
+    return null;
+  }
+
+  let masterworkInfo: DimMasterwork | null = null;
+
+  // Pre-Forsaken Masterwork
+  if (createdItem.masterwork) {
+    masterworkInfo = buildMasterworkInfo(createdItem.sockets, defs);
+    if (masterworkInfo !== null) {
+      console.log('pre-forsaken', createdItem.name);
+    }
+  }
+
+  // Forsaken Masterwork
+  if (!masterworkInfo) {
+    masterworkInfo = buildForsakenMasterworkInfo(createdItem, defs);
+  }
+
+  return masterworkInfo;
+}
+
+/**
+ * Post-Forsaken weapons store their masterwork info and kill tracker on different plugs.
+ */
+function buildForsakenMasterworkInfo(
+  createdItem: D2Item,
+  defs: D2ManifestDefinitions
+): DimMasterwork | null {
+  const masterworkSocket = createdItem.sockets!.sockets.find((socket) =>
+    Boolean(
+      socket.plug &&
+        socket.plug.plugItem.plug &&
+        (socket.plug.plugItem.plug.plugCategoryIdentifier.includes('masterworks.stat') ||
+          socket.plug.plugItem.plug.plugCategoryIdentifier.endsWith('_masterwork'))
+    )
+  );
+  if (masterworkSocket && masterworkSocket.plug) {
+    if (masterworkSocket.plug.plugItem.investmentStats.length) {
+      const masterwork = masterworkSocket.plug.plugItem.investmentStats[0];
+      if (createdItem.bucket && createdItem.bucket.sort === 'Armor') {
+        createdItem.dmg = [null, 'heroic', 'arc', 'solar', 'void'][
+          resistanceMods[masterwork.statTypeHash]
+        ] as typeof createdItem.dmg;
+      }
+
+      if (
+        (createdItem.bucket.sort === 'Armor' && masterwork.value === 5) ||
+        (createdItem.bucket.sort === 'Weapon' && masterwork.value === 10)
+      ) {
+        createdItem.masterwork = true;
+      }
+      const statDef = defs.Stat.get(masterwork.statTypeHash);
+
+      createdItem.masterworkInfo = {
+        typeName: null,
+        typeIcon: masterworkSocket.plug.plugItem.displayProperties.icon,
+        typeDesc: masterworkSocket.plug.plugItem.displayProperties.description,
+        statHash: masterwork.statTypeHash,
+        statName: statDef.displayProperties.name,
+        statValue: masterwork.value
+      };
+    }
+
+    const killTracker = createdItem.sockets!.sockets.find((socket) =>
+      Boolean(idx(socket.plug, (p) => p.plugObjectives.length))
+    );
+
+    if (
+      killTracker &&
+      killTracker.plug &&
+      killTracker.plug.plugObjectives &&
+      killTracker.plug.plugObjectives.length
+    ) {
+      const plugObjective = killTracker.plug.plugObjectives[0];
+
+      const objectiveDef = defs.Objective.get(plugObjective.objectiveHash);
+
+      console.log('Forsaken masterwork', createdItem.name);
+      return {
+        ...createdItem.masterworkInfo,
+        progress: plugObjective.progress,
+        typeIcon: objectiveDef.displayProperties.icon,
+        typeDesc: objectiveDef.progressDescription,
+        typeName: [3244015567, 2285636663, 38912240].includes(killTracker.plug.plugItem.hash)
+          ? 'Crucible'
+          : 'Vanguard'
+      };
+    }
+  }
+
+  return null;
+}
+
+/**
+ * Pre-Forsaken weapons store their masterwork info on an objective of a plug.
+ */
+function buildMasterworkInfo(
+  sockets: DimSockets,
+  defs: D2ManifestDefinitions
+): DimMasterwork | null {
+  const socket = sockets.sockets.find((socket) =>
+    Boolean(socket.plug && socket.plug.plugObjectives.length)
+  );
+  if (
+    !socket ||
+    !socket.plug ||
+    !socket.plug.plugObjectives ||
+    !socket.plug.plugObjectives.length
+  ) {
+    return null;
+  }
+  const plugObjective = socket.plug.plugObjectives[0];
+  const investmentStats = socket.plug.plugItem.investmentStats;
+  if (!investmentStats || !investmentStats.length) {
+    return null;
+  }
+  const statHash = investmentStats[0].statTypeHash;
+
+  const objectiveDef = defs.Objective.get(plugObjective.objectiveHash);
+  const statDef = defs.Stat.get(statHash);
+
+  if (!objectiveDef || !statDef) {
+    return null;
+  }
+
+  return {
+    progress: plugObjective.progress,
+    typeName: socket.plug.plugItem.plug.plugCategoryHash === 2109207426 ? 'Vanguard' : 'Crucible',
+    typeIcon: objectiveDef.displayProperties.icon,
+    typeDesc: objectiveDef.progressDescription,
+    statHash,
+    statName: statDef.displayProperties.name,
+    statValue: investmentStats[0].value
+  };
+}
